decouple HuggingFace LLM provider from LangChain dependency
