+++ docs/source/build-workflows/llms/index.md
@@ -29,6 +29,7 @@ NVIDIA NeMo Agent toolkit supports the following LLM providers:
 | [AWS Bedrock](https://aws.amazon.com/bedrock/) | `aws_bedrock` | AWS Bedrock API |
 | [Azure OpenAI](https://learn.microsoft.com/en-us/azure/ai-foundry/openai/quickstart) | `azure_openai` | Azure OpenAI API |
 | [LiteLLM](https://github.com/BerriAI/litellm) | `litellm` | LiteLLM API |
+| [HuggingFace](https://huggingface.co) | `huggingface` | HuggingFace API |
 
 
 ## LLM Configuration
@@ -53,6 +54,9 @@ llms:
   litellm_llm:
     _type: litellm
     model_name: gpt-4o
+  huggingface_llm:
+    _type: huggingface
+    model_name: Qwen/Qwen3Guard-Gen-0.6B
 ```
 
 ### NVIDIA NIM
@@ -148,6 +152,26 @@ The LiteLLM LLM provider is defined by the {py:class}`~nat.llm.litellm_llm.LiteL
 * `temperature` - The temperature to use for the model
 * `top_p` - The top-p value to use for the model
 * `max_retries` - The maximum number of retries for the request
+
+### HuggingFace
+
+HuggingFace is a general purpose LLM provider that can be used with any model supported by the HuggingFace API.
+See the [HuggingFace documentation](https://huggingface.co/docs) for more information.
+
+**Installation:**
+
+```bash
+uv pip install "nvidia-nat[huggingface]"
+```
+
+The HuggingFace LLM provider is defined by the {py:class}`~nat.llm.huggingface_llm.HuggingFaceConfig` class.
+
+* `model_name` - The HuggingFace model name or path (for example, `Qwen/Qwen3Guard-Gen-0.6B`)
+* `device` - Device for model execution: `cpu`, `cuda`, `cuda:0`, or `auto` (default: `auto`)
+* `torch_dtype` - Torch data type: `float16`, `bfloat16`, `float32`, or `auto` (default: `auto`)
+* `max_new_tokens` - Maximum number of new tokens to generate (default: `128`)
+* `temperature` - Sampling temperature (default: `0.0`)
+* `trust_remote_code` - Whether to trust remote code when loading the model (default: `false`)
 
 ### NVIDIA Dynamo (experimental)
