+++ pyproject.toml
@@ -96,7 +96,7 @@ agno = ["nvidia-nat-agno"]
 autogen = ["nvidia-nat-autogen"]
 crewai = ["nvidia-nat-crewai"]
 data-flywheel = ["nvidia-nat-data-flywheel"]
-huggingface = ["transformers", "torch", "accelerate"]
+huggingface = ["nvidia-nat-langchain", "transformers", "torch", "accelerate"]
 ingestion = ["nvidia-nat-ingestion"] # meta-package
 langchain = ["nvidia-nat-langchain"]
 llama-index = ["nvidia-nat-llama-index"]
