+++ src/nat/llm/huggingface_llm.py
@@ -12,23 +12,12 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-"""
-HuggingFace Transformers LLM Provider - Local in-process model execution.
-"""
+"""HuggingFace Transformers LLM Provider - Local in-process model execution."""
 
 import logging
 from collections.abc import AsyncIterator
 from typing import Any
 
-from langchain_core.callbacks import AsyncCallbackManagerForLLMRun
-from langchain_core.callbacks import CallbackManagerForLLMRun
-from langchain_core.language_models import BaseChatModel
-from langchain_core.messages import AIMessage
-from langchain_core.messages import AIMessageChunk
-from langchain_core.messages import BaseMessage
-from langchain_core.outputs import ChatGeneration
-from langchain_core.outputs import ChatResult
-from pydantic import ConfigDict
 from pydantic import Field
 
 from nat.builder.builder import Builder
@@ -43,7 +32,7 @@ logger = logging.getLogger(__name__)
 # - During nat serve: Cached while server runs, cleaned up on shutdown
 # - During nat red-team: Cached across all evaluation queries, cleaned up when complete
 # - During nat run: Cached for single workflow execution, cleaned up when done
-_model_cache = {}
+_model_cache: dict[str, dict[str, Any]] = {}
 
 
 class HuggingFaceConfig(LLMBaseConfig, name="huggingface"):
@@ -63,253 +52,9 @@ class HuggingFaceConfig(LLMBaseConfig, name="huggingface"):
     trust_remote_code: bool = Field(default=False, description="Trust remote code when loading model")
 
 
-class HuggingFaceModel(BaseChatModel):
-    """LangChain-compatible wrapper for local HuggingFace models.
-
-    This class inherits from BaseChatModel to provide proper LangChain integration
-    for locally loaded HuggingFace Transformers models.
-    """
-
-    model_config = ConfigDict(arbitrary_types_allowed=True)
-
-    # Attributes (set during initialization)
-    _model_name: str
-    _config: HuggingFaceConfig
-    _model: Any
-    _tokenizer: Any
-    _torch: Any
-
-    def __init__(self, model_name: str, config: HuggingFaceConfig):
-        """Initialize HuggingFace model wrapper.
-
-        Args:
-            model_name: Name of the loaded model
-            config: Configuration for the model
-        """
-        # Get from cache
-        if model_name not in _model_cache:
-            raise ValueError(f"Model {model_name} not loaded in cache")
-
-        cached = _model_cache[model_name]
-
-        # Initialize parent
-        super().__init__()
-
-        # Set private attributes
-        self._model_name = model_name
-        self._config = config
-        self._model = cached["model"]
-        self._tokenizer = cached["tokenizer"]
-        self._torch = cached["torch"]
-
-    @property
-    def model_name(self) -> str:
-        """Return the model name."""
-        return self._model_name
-
-    @property
-    def config(self) -> HuggingFaceConfig:
-        """Return the model configuration."""
-        return self._config
-
-    @property
-    def model(self):
-        """Return the HuggingFace model."""
-        return self._model
-
-    @property
-    def tokenizer(self):
-        """Return the tokenizer."""
-        return self._tokenizer
-
-    @property
-    def torch(self):
-        """Return the torch module."""
-        return self._torch
-
-    @property
-    def _llm_type(self) -> str:
-        """Return identifier for the LLM type."""
-        return "huggingface"
-
-    def _prepare_text(self, messages: list[BaseMessage] | list[dict] | str) -> str:
-        """Convert messages to text using chat template or fallback.
-
-        Args:
-            messages: Input messages in various formats (BaseMessage list, dict list, or string)
-
-        Returns:
-            Formatted text string ready for tokenization
-        """
-        # Convert BaseMessage objects to dict format for template
-        if isinstance(messages, list) and len(messages) > 0:
-            # Handle LangChain BaseMessage objects
-            if hasattr(messages[0], "type") and hasattr(messages[0], "content"):
-                messages = [{
-                    "role": msg.type, "content": msg.content
-                } for msg in messages]  # type: ignore[attr-defined]
-
-            # Try using chat template
-            try:
-                text = self.tokenizer.apply_chat_template(messages, tokenize=False, add_generation_prompt=True)
-            except (ValueError, TypeError, KeyError, AttributeError) as e:
-                # Fallback: just use the last message content
-                logger.debug("Chat template application failed: %s, using fallback", e)
-                last_msg = messages[-1]
-                text = last_msg.get("content", str(last_msg)) if isinstance(last_msg, dict) else str(last_msg)
-        else:
-            text = str(messages)
-        return text
-
-    def _generate(
-        self,
-        messages: list[BaseMessage],
-        stop: list[str] | None = None,
-        run_manager: CallbackManagerForLLMRun | None = None,
-        **kwargs: Any,
-    ) -> ChatResult:
-        """Generate response synchronously (required by BaseChatModel).
-
-        Args:
-            messages: List of message objects.
-            stop: Optional list of stop sequences.
-            run_manager: Optional callback manager.
-            kwargs: Additional generation parameters.
-
-        Returns:
-            ChatResult containing the generated response.
-        """
-        # Wrap async implementation
-        import asyncio
-        try:
-            loop = asyncio.get_event_loop()
-        except RuntimeError:
-            loop = asyncio.new_event_loop()
-            asyncio.set_event_loop(loop)
-
-        # Note: run_manager is sync but _agenerate expects async, so we don't pass it
-        result = loop.run_until_complete(self._agenerate(messages, stop=stop, **kwargs))
-        return result
-
-    async def _agenerate(
-        self,
-        messages: list[BaseMessage],
-        stop: list[str] | None = None,
-        run_manager: AsyncCallbackManagerForLLMRun | None = None,
-        **kwargs: Any,
-    ) -> ChatResult:
-        """Generate response asynchronously (called by BaseChatModel.ainvoke).
-
-        Args:
-            messages: List of message objects.
-            stop: Optional list of stop sequences.
-            run_manager: Optional callback manager.
-            kwargs: Additional generation parameters.
-
-        Returns:
-            ChatResult containing the generated response.
-        """
-        # Convert messages to text
-        text = self._prepare_text(messages)
-
-        # Tokenize
-        model_inputs = self.tokenizer([text], return_tensors="pt").to(self.model.device)
-
-        # Generate
-        with self.torch.no_grad():
-            generated_ids = self.model.generate(**model_inputs,
-                                                max_new_tokens=self.config.max_new_tokens,
-                                                temperature=self.config.temperature
-                                                if self.config.temperature > 0 else None,
-                                                do_sample=self.config.temperature > 0,
-                                                pad_token_id=self.tokenizer.eos_token_id)
-
-        # Decode (only new tokens)
-        output_ids = generated_ids[0][len(model_inputs.input_ids[0]):].tolist()
-        content = self.tokenizer.decode(output_ids, skip_special_tokens=True)
-
-        # Return ChatResult (BaseChatModel format)
-        message = AIMessage(content=content)
-        generation = ChatGeneration(message=message)
-        return ChatResult(generations=[generation])
-
-    async def _astream(
-        self,
-        messages: list[BaseMessage],
-        stop: list[str] | None = None,
-        run_manager: AsyncCallbackManagerForLLMRun | None = None,
-        **kwargs: Any,
-    ):
-        """Stream response tokens as they are generated (called by BaseChatModel.astream).
-
-        Args:
-            messages: List of message objects.
-            stop: Optional list of stop sequences.
-            run_manager: Optional callback manager.
-            kwargs: Additional generation parameters.
-
-        Yields:
-            ChatGenerationChunk objects containing token chunks.
-        """
-        from langchain_core.outputs import ChatGenerationChunk
-
-        try:
-            from transformers import TextIteratorStreamer
-        except ImportError:
-            # Fallback: if TextIteratorStreamer not available, yield full response
-            logger.debug("TextIteratorStreamer not available, falling back to non-streaming")
-            result = await self._agenerate(messages, stop=stop, run_manager=run_manager, **kwargs)
-            # Convert AIMessage to AIMessageChunk for streaming
-            full_message = result.generations[0].message
-            chunk = AIMessageChunk(content=full_message.content)
-            yield ChatGenerationChunk(message=chunk)
-            return
-
-        # Convert messages to text
-        text = self._prepare_text(messages)
-        model_inputs = self.tokenizer([text], return_tensors="pt").to(self.model.device)
-
-        # Create streamer for token-by-token generation
-        streamer = TextIteratorStreamer(self.tokenizer, skip_special_tokens=True, skip_prompt=True)
-
-        # Prepare generation kwargs
-        generation_kwargs = {
-            **model_inputs,
-            "streamer": streamer,
-            "max_new_tokens": self.config.max_new_tokens,
-            "temperature": self.config.temperature if self.config.temperature > 0 else None,
-            "do_sample": self.config.temperature > 0,
-            "pad_token_id": self.tokenizer.eos_token_id
-        }
-
-        # Start generation in background thread (model.generate is blocking)
-        import asyncio
-        import threading
-        thread = threading.Thread(target=self.model.generate, kwargs=generation_kwargs)
-        thread.start()
-
-        # Stream tokens as they're generated
-        try:
-            for token_text in streamer:
-                # Yield control to event loop
-                await asyncio.sleep(0)
-
-                # Return chunk in BaseChatModel format
-                chunk = AIMessageChunk(content=token_text)
-                yield ChatGenerationChunk(message=chunk)
-        finally:
-            # Ensure thread completes
-            thread.join()
-
-    def bind_tools(self, tools, **kwargs):
-        """Bind tools to the LLM. Returns self to maintain fluent interface."""
-        # HuggingFace models don't support tool calling, but we return self for compatibility
-        return self
-
-    def bind(self, **kwargs):
-        """Bind additional parameters to the LLM. Returns self to maintain fluent interface."""
-        # HuggingFace models don't support parameter binding, but we return self for compatibility
-        return self
+def get_cached_model(model_name: str) -> dict[str, Any] | None:
+    """Return cached model data (model, tokenizer, torch) or None if not loaded."""
+    return _model_cache.get(model_name)
 
 
 async def _cleanup_model(model_name: str) -> None:
@@ -386,18 +131,3 @@ async def huggingface_provider(
     finally:
         # Cleanup when workflow/application shuts down
         await _cleanup_model(config.model_name)
-
-
-def get_huggingface_model(model_name: str, config: HuggingFaceConfig):
-    """Create a HuggingFace model wrapper for a loaded model.
-
-    Args:
-        model_name: Name of the model to retrieve.
-        config: Configuration for the model wrapper.
-
-    Returns:
-        HuggingFaceModel instance or None if model not loaded.
-    """
-    if model_name in _model_cache:
-        return HuggingFaceModel(model_name, config)
-    return None
